task-1/airflow/requirements.txt
apache-airflow-providers-postgres==5.12.0
apache-airflow-providers-google==10.18.0
apache-airflow-providers-amazon==8.22.0
apache-airflow-providers-apache-spark==4.6.1
apache-airflow-providers-apache-kafka==1.4.0
pandas==2.2.2
psycopg2-binary==2.9.9